\name{ethnicity}
\alias{ethnicity}
\docType{data}
\title{
Toy example using New Zealand level 1 ethnicity values. The statistical standard for collecting ethnicity data requires that respondents can mark all that are applicable.  The level 1 values are "Māori", "Pacific Peoples" (ie, Pacific Island ethnicities), "Asian", "European", and "MELAA" (Middle Eastern, Latin American, and African).
}
\description{
Artficial data
}
\usage{data("ethnicity")}
\format{
 An object of class \code{mr}
}

\examples{
data(ethnicity)
ethnicity
}
\keyword{datasets}
